rename(output): rename output.js to index.js
